moved spec hash annotation logic to be set after deployment.
