+++ internal/controller/pipelinerollout_controller_test.go
@@ -331,7 +331,7 @@ func Test_makeChildResourceFromRolloutAndUpdateSpecHash_InvalidType(t *testing.T
 
 	invalidType := kubernetes.GenericObject{}
 
-	_, _, err := makeChildResourceFromRolloutAndUpdateSpecHash(ctx, restConfig, &invalidType)
+	_, _, _, err := makeChildResourceFromRolloutAndUpdateSpecHash(ctx, restConfig, &invalidType)
 
 	assert.Error(t, err)
 	assert.Equal(t, "invalid rollout type", err.Error())
@@ -352,48 +352,6 @@ func Test_makeChildResourceFromRolloutAndUpdateSpecHash_PipelineRollout_Unmarsha
 		},
 	}
 
-	_, _, err := makeChildResourceFromRolloutAndUpdateSpecHash(ctx, restConfig, pipelineRolloutInvalidSpec)
+	_, _, _, err := makeChildResourceFromRolloutAndUpdateSpecHash(ctx, restConfig, pipelineRolloutInvalidSpec)
 	assert.Error(t, err)
-}
-
-func Test_setAnnotation(t *testing.T) {
-	key1, value1 := "some_key_1", "some_value_1"
-	key2, value2 := "some_key_2", "some_value_2"
-
-	pipelineRollout := &apiv1.PipelineRollout{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "pipeline-test",
-			Namespace: "default",
-		},
-		Spec: apiv1.PipelineRolloutSpec{
-			Pipeline: runtime.RawExtension{Raw: []byte(`{"key":"value"}`)},
-		},
-	}
-
-	// Invoke the method under test
-	setAnnotation(pipelineRollout, key1, value1)
-
-	// Check if the annotation was correctly set
-	annotations := pipelineRollout.GetAnnotations()
-	assert.NotNil(t, annotations, "Expected annotations to be set on the pipelineRollout object")
-	assert.Contains(t, annotations, key1, "Expected the key to be set in the annotations")
-	assert.Equal(t, value1, annotations[key1], "Expected the value to be set for the key in the annotations")
-
-	// Overwrite existing annotation
-	newValue := "new_value"
-	setAnnotation(pipelineRollout, key1, newValue)
-
-	// Check if the annotation was correctly updated
-	annotations = pipelineRollout.GetAnnotations()
-	assert.NotNil(t, annotations, "Expected annotations to be set on the pipelineRollout object")
-	assert.Contains(t, annotations, key1, "Expected the key to be set in the annotations")
-	assert.NotEqual(t, value1, annotations[key1], "Expected the old value to be replaced")
-	assert.Equal(t, newValue, annotations[key1], "Expected the new value to be set for the key in the annotations")
-
-	// Add one more annotation
-	setAnnotation(pipelineRollout, key2, value2)
-	assert.NotNil(t, annotations, "Expected annotations to be set on the pipelineRollout object")
-	assert.Len(t, annotations, 2, "Expected annotations to be of length 2")
-	assert.Contains(t, annotations, key2, "Expected the key to be set in the annotations")
-	assert.Equal(t, value2, annotations[key2], "Expected the value to be set for the key in the annotations")
 }
